RELEASE Setting up week 5 weekly release

+++ backend/scripts/dbScripts.js
@@ -32,18 +32,18 @@ const dbScripts = {
         DROP TABLE IF EXISTS TASKSBASIC, TASKSADVANCED;
 
         CREATE TABLE IF NOT EXISTS TASKSBASIC(
-            taskId INT UNIQUE NOT NULL,
+            taskId BIGINT UNIQUE NOT NULL,
             dueDate DATE NOT NULL,
             dueTime TIME NOT NULL,
             duration INT NOT NULL,
-            projectId INT NOT NULL,
+            projectId BIGINT NOT NULL,
             PRIMARY KEY (taskId)
         );
 
         CREATE TABLE IF NOT EXISTS TASKSADVANCED(
-            taskId INT UNIQUE NOT NULL,
+            taskId BIGINT UNIQUE NOT NULL,
             duration INT NOT NULL,
-            projectId INT NOT NULL,
+            projectId BIGINT NOT NULL,
             PRIMARY KEY (taskId)
         );
         `;
